requirements.txt
streamlit
langchain
langchain-openai
langchain-community
tavily-python
python-dotenv